update appium lazyLoad + helper lib tests with retry mechanism (unstable test)
