+++ eyes.appium.java/src/test/java/com/applitools/eyes/appium/android/AndroidLazyLoadTest.java
@@ -2,16 +2,8 @@ package com.applitools.eyes.appium.android;
 
 import com.applitools.eyes.LazyLoadOptions;
 import com.applitools.eyes.appium.Target;
-import io.appium.java_client.AppiumBy;
-import org.openqa.selenium.By;
-import org.openqa.selenium.WebElement;
-import org.openqa.selenium.interactions.PointerInput;
-import org.openqa.selenium.interactions.Sequence;
-import org.testng.Assert;
 import org.testng.annotations.Test;
 
-import java.time.Duration;
-import java.util.Arrays;
 
 public class AndroidLazyLoadTest extends AndroidTestSetup {
 
@@ -19,63 +11,25 @@ public class AndroidLazyLoadTest extends AndroidTestSetup {
     public static final int PIXEL_3XL_OFFSET = 13601;
 
     @Test
-    public void testHelperLibOffsetCalculation() throws InterruptedException {
-        gestureScrollAndClick();
-        String fieldCommand = "offset_async;recycler_view;0;0;0;5000";
+    public void testLazyLoad() {
+        testLazyLoadWithRetry();
+    }
 
-        WebElement edt = null;
-        for (int i = 0; i < 3; i++) {
+    private void testLazyLoadWithRetry() {
+        for (int j = 0; j < 2; j++) {
             try {
-                Thread.sleep(50);
-                edt = driver.findElement(By.xpath("//*[@content-desc=\"EyesAppiumHelperEDT\"]"));
+                gestureScrollAndClick();
+                eyes.open(driver, getApplicationName(), "Check LazyLoad");
+                LazyLoadOptions lazyLoadOptions = new LazyLoadOptions().waitingTime(-PIXEL_5_OFFSET);
+                eyes.check(Target.window().fully().lazyLoad(lazyLoadOptions).withName("lazyLoad"));
+                eyes.close();
                 break;
             } catch (Exception e) {
-                if(++i >= 3)
+                driver.navigate().back();
+                if(++j >= 2)
                     throw e;
             }
         }
-
-        edt.sendKeys(fieldCommand);
-        edt.click();
-
-        do { Thread.sleep(1000); }
-        while (edt.getText().equals("WAIT"));
-
-        String value = edt.getText().split(";", 2)[0];
-        driver.navigate().back();
-
-        Assert.assertEquals(Integer.parseInt(value), PIXEL_5_OFFSET);
-
-    }
-
-    @Test
-    public void testLazyLoad() {
-        gestureScrollAndClick();
-        eyes.open(driver, getApplicationName(), "Check LazyLoad");
-        LazyLoadOptions lazyLoadOptions = new LazyLoadOptions().waitingTime(-PIXEL_5_OFFSET);
-        eyes.check(Target.window().fully().lazyLoad(lazyLoadOptions).withName("lazyLoad"));
-        eyes.close();
-    }
-
-    private void gestureScrollAndClick() {
-        // create new pointer action
-        PointerInput finger = new PointerInput(PointerInput.Kind.TOUCH, "finger");
-        // create new sequence of actions
-        Sequence dragNDrop = new Sequence(finger, 1);
-
-        // add pointer movement
-        dragNDrop.addAction(finger.createPointerMove(Duration.ofMillis(0),
-                PointerInput.Origin.viewport(), 5, 1000));
-        // pointer down - left click
-        dragNDrop.addAction(finger.createPointerDown(PointerInput.MouseButton.LEFT.asArg()));
-        // add pointer movement
-        dragNDrop.addAction(finger.createPointerMove(Duration.ofMillis(700),
-                PointerInput.Origin.viewport(),5, 100));
-        // pointer up - release left click
-        dragNDrop.addAction(finger.createPointerUp(PointerInput.MouseButton.LEFT.asArg()));
-        driver.perform(Arrays.asList(dragNDrop));
-
-        driver.findElement(AppiumBy.id("btn_large_recyclerView_activity")).click();
     }
 
 }
