+++ eyes.appium.java/src/test/java/com/applitools/eyes/appium/android/AndroidTestSetup.java
@@ -1,10 +1,15 @@
 package com.applitools.eyes.appium.android;
 
 import com.applitools.eyes.appium.TestSetup;
+import io.appium.java_client.AppiumBy;
 import io.appium.java_client.android.AndroidDriver;
+import org.openqa.selenium.interactions.PointerInput;
+import org.openqa.selenium.interactions.Sequence;
 
 import java.net.MalformedURLException;
 import java.net.URL;
+import java.time.Duration;
+import java.util.Arrays;
 
 public abstract class AndroidTestSetup extends TestSetup {
 
@@ -39,5 +44,26 @@ public abstract class AndroidTestSetup extends TestSetup {
     @Override
     protected String getApplicationName() {
         return "Java Appium - Android";
+    }
+
+    public void gestureScrollAndClick() {
+        // create new pointer action
+        PointerInput finger = new PointerInput(PointerInput.Kind.TOUCH, "finger");
+        // create new sequence of actions
+        Sequence dragNDrop = new Sequence(finger, 1);
+
+        // add pointer movement
+        dragNDrop.addAction(finger.createPointerMove(Duration.ofMillis(0),
+                PointerInput.Origin.viewport(), 5, 1000));
+        // pointer down - left click
+        dragNDrop.addAction(finger.createPointerDown(PointerInput.MouseButton.LEFT.asArg()));
+        // add pointer movement
+        dragNDrop.addAction(finger.createPointerMove(Duration.ofMillis(700),
+                PointerInput.Origin.viewport(),5, 100));
+        // pointer up - release left click
+        dragNDrop.addAction(finger.createPointerUp(PointerInput.MouseButton.LEFT.asArg()));
+        driver.perform(Arrays.asList(dragNDrop));
+
+        driver.findElement(AppiumBy.id("btn_large_recyclerView_activity")).click();
     }
 }
